requirements.txt
pandas
usaddress
psycopg2-binary
rapidfuzz
tqdm
phonetics
pytest
fastapi
uvicorn
psutil
